Allows boomerang unstake
